refactor(format): extract markdown derived renderer
